Move SetPrefix/SetId to J interface, restore template-specific TreeHelper
- Add SetPrefix<T> and SetId<T> as static methods on partial interface J
  in OpenRewrite.Java, where recipe authors would naturally find them
- Restore TreeHelper in OpenRewrite.CSharp.Template as internal, containing
  only structural comparison helpers used by PatternMatchingComparator
- Update TemplateEngine to call J.SetPrefix/J.SetId
